+++ src/connectors/reader/toolbar.js
@@ -2,13 +2,19 @@ import { useDispatch, useSelector } from 'react-redux'
 import { ReaderNav } from 'components/reader/nav'
 
 import { setColorMode } from 'connectors/app/app.state'
-import { toggleSidebar } from 'containers/read/read.state'
-import { favoriteItem } from 'containers/read/read.state'
-import { unFavoriteItem } from 'containers/read/read.state'
-import { archiveItem } from 'containers/read/read.state'
-import { unArchiveItem } from 'containers/read/read.state'
+import { toggleSidebar } from 'containers/read/reader.state'
+
+import { mutationFavorite } from 'connectors/items/mutation-favorite.state'
+import { mutationUnFavorite } from 'connectors/items/mutation-favorite.state'
+import { mutationArchive } from 'connectors/items/mutation-archive.state'
 import { mutationDelete } from 'connectors/items/mutation-delete.state'
+import { mutationUnArchive } from 'connectors/items/mutation-archive.state'
 import { mutationTagItem } from 'connectors/items/mutation-tagging.state'
+
+import { updateLineHeight } from 'containers/read/reader-settings.state'
+import { updateColumnWidth } from 'containers/read/reader-settings.state'
+import { updateFontSize } from 'containers/read/reader-settings.state'
+import { updateFontType } from 'containers/read/reader-settings.state'
 
 import { shareAction } from 'connectors/share-modal/share-modal.state'
 
@@ -17,26 +23,22 @@ import { sendSnowplowEvent } from 'connectors/snowplow/snowplow.state'
 export const Toolbar = ({ id }) => {
   const dispatch = useDispatch()
 
-  const item = useSelector((state) => state.reader.articleItem)
-  const savedData = useSelector((state) => state.reader.savedData)
-  const sideBarOpen = useSelector((state) => state.reader.sideBarOpen)
+  const item = useSelector((state) => state.items[id])
+  const savedData = useSelector((state) => state.itemsSaved[id])
+  const sideBarOpen = useSelector((state) => state.readerGraph.sideBarOpen)
 
   const isPremium = useSelector((state) => state.user.premium_status === '1')
   const colorMode = useSelector((state) => state?.app?.colorMode)
 
   // Display settings
-  const lineHeight = useSelector((state) => state.reader.lineHeight)
-  const columnWidth = useSelector((state) => state.reader.columnWidth)
-  const fontSize = useSelector((state) => state.reader.fontSize)
-  const fontFamily = useSelector((state) => state.reader.fontFamily)
+  const lineHeight = useSelector((state) => state.readerSettings.lineHeight)
+  const columnWidth = useSelector((state) => state.readerSettings.columnWidth)
+  const fontSize = useSelector((state) => state.readerSettings.fontSize)
+  const fontFamily = useSelector((state) => state.readerSettings.fontFamily)
 
   const { analyticsData } = item
 
-  const {
-    isArchived,
-    isFavorite,
-    tags
-  } = savedData
+  const { isArchived, isFavorite, tags } = savedData
 
   const handleSidebarToggle = () => dispatch(toggleSidebar())
 
@@ -56,14 +58,14 @@ export const Toolbar = ({ id }) => {
   }
 
   const toggleFavorite = () => {
-    const favoriteAction = isFavorite ? unFavoriteItem : favoriteItem
+    const favoriteAction = isFavorite ? mutationUnFavorite : mutationFavorite
     const identifier = isFavorite ? 'reader.un-favorite' : 'reader.favorite'
     dispatch(sendSnowplowEvent(identifier, analyticsData))
     dispatch(favoriteAction(id))
   }
 
   const itemArchive = () => {
-    const archiveAction = isArchived ? unArchiveItem : archiveItem
+    const archiveAction = isArchived ? mutationUnArchive : mutationArchive
     const identifier = isArchived ? 'reader.un-archive' : 'reader.archive'
     dispatch(sendSnowplowEvent(identifier, analyticsData))
     dispatch(archiveAction(id))
@@ -91,6 +93,10 @@ export const Toolbar = ({ id }) => {
       onVisible={handleImpression}
       colorMode={colorMode}
       setColorMode={setAppColorMode}
+      updateLineHeight={updateLineHeight}
+      updateColumnWidth={updateColumnWidth}
+      updateFontSize={updateFontSize}
+      updateFontType={updateFontType}
     />
   )
 }